+++ libs/designer/src/lib/ui/mcp/panel/connector/ConnectorPanelInner.tsx
