+++ libs/designer/src/lib/ui/mcp/panel/connector/tabs/connectorsTab.tsx
