+++ libs/designer/src/lib/ui/mcp/panel/connector/tabs/operationsTab.tsx
