+++ libs/designer/src/lib/ui/mcp/panel/connector/usePanelTabs.tsx
