+++ libs/designer/src/lib/ui/mcp/panel/mcpPanelRoot.tsx
@@ -4,9 +4,9 @@ import type { AppDispatch } from '../../../';
 import { closePanel, McpPanelView } from '../../../core/state/mcp/panel/mcpPanelSlice';
 import { useMcpPanelStyles } from './styles';
 import type { RootState } from '../../../core/state/mcp/store';
-import { ConnectorPanelInner } from './connectorPanel/ConnectorPanelInner';
+import { ConnectorPanelInner } from './connector/ConnectorPanelInner';
 import { Drawer } from '@fluentui/react-components';
-import { EditOperationPanelInner } from './editOperationPanel/EditOperationPanelInner';
+import { EditOperationPanelInner } from './operation/EditOperationPanelInner';
 
 const MIN_WIDTH = 300;
 const MAX_WIDTH = 1200;
