+++ libs/designer/src/lib/ui/mcp/panel/operation/EditOperationPanelInner.tsx
